requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
email_validator==2.1.0.post1
python-dotenv==1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.8.0
beautifulsoup4
python-telegram-bot>=20.6
psycopg2-binary
gunicorn